requirements.txt
pydantic==2.1.1
requests==2.31.0
azure.functions
azure-functions-worker
